# components/org.wso2.carbon.stream.processor.core/src/test/resources/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Siddhi App Service
  description: This provides information about the REST APIs available for WSO2 Siddhi.
paths:
  /siddhi-apps:
    post:
      tags:
        - artifact
      summary: Saves the Siddhi App
      description: |
        Saves the Siddhi App.
        Request "**siddhiApp** explains the Siddhi Query.
      consumes:
        - text/plain
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Siddhi App
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The Siddhi App is successfully validated and saved.
        '400':
          description: A validation error occured.
        '409':
          description: A Siddhi App with the given name already exists.
        '500':
          description: An unexpected error occured.
    put:
      tags:
        - artifact
      summary: Updates the Siddhi App.
      description: |
        Updates the Siddhi App.
        Request "**siddhiApp** explains the Siddhi Query.
      consumes:
        - text/plain
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Siddhi App
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The Siddhi App is successfully validated and updated.
        '201':
          description: The Siddhi App is successfully validated and saved.
        '400':
          description: A validation error occured.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
        '500':
          description: An unexpected error occured.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
    get:
      tags:
        - artifact
      summary: Lists Siddhi Apps
      description: |
        Provides the name list of Siddhi Apps that exist.
      produces:
        - application/json
      parameters:
        - name: isActive
          in: query
          description: Retrieves only active/inactive Siddhi Apps as specified.
          required: false
          type: boolean
      responses:
        '200':
          description: The Siddhi Apps are successfully retrieved.
  '/siddhi-apps/{appName}':
    delete:
      tags:
        - artifact
      summary: Deletes the Siddhi App.
      description: |
        Removes the currently deployed Siddhi App with the name specified.
        Path param of **appName** determines name of the Siddhi app.
      produces:
        - application/json
      parameters:
        - name: appName
          in: path
          description: The name of the Siddhi App.
          required: true
          type: string
      responses:
        '200':
          description: The Siddhi App is successfully deleted.
        '400':
          description: The Siddhi App name provided is invalid.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
        '404':
          description: The Siddhi App is not found.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
        '500':
          description: An unexpected error occured.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
    get:
      tags:
        - artifact
      summary: Retrieves the specified Siddhi App.
      description: Retrieves the Siddhi App with the specified name.
      produces:
        - application/json
      parameters:
        - name: appName
          in: path
          description: The name of the Siddhi App.
          required: true
          type: string
      responses:
        '200':
          description: The Siddhi App is successfully retrieved.
          schema:
            $ref: '#/definitions/SiddhiAppContent'
        '404':
          description: The Siddhi App specified is not found.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
  '/siddhi-apps/{appName}/status':
    get:
      tags:
        - artifact
      summary: Fetches the status of the Siddhi App.
      description: |
        Fetches the status of the Siddhi App.
      produces:
        - application/json
      parameters:
        - name: appName
          in: path
          description: The name of the Siddhi App.
          required: true
          type: string
      responses:
        '200':
          description: The state of the Siddhi App is successfully retrieved.
          schema:
            $ref: '#/definitions/SiddhiAppStatus'
        '404':
          description: The Siddhi App specified is not found.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
  '/siddhi-apps/{appName}/backup':
    post:
      tags:
        - state
      summary: Backs up the current state of a Siddhi App.
      description: |
        Backs up the current state of the specified Siddhi App.
      produces:
        - application/json
      parameters:
        - name: appName
          in: path
          description: The name of the Siddhi App.
          required: true
          type: string
      responses:
        '200':
          description: The state of the Siddhi App is successfully persisted.
          schema:
            $ref: '#/definitions/SiddhiAppRevision'
        '404':
          description: The Siddhi App specified is not found.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
        '500':
          description: An unexpected error occured.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
  '/siddhi-apps/{appName}/restore':
    post:
      tags:
        - state
      summary: Restores the state of a Siddhi App.
      description: |
        Restores the state of the specified Siddhi App.
      produces:
        - application/json
      parameters:
        - name: appName
          in: path
          description: The name of the Siddhi App.
          required: true
          type: string
        - name: revision
          in: query
          description: The revision number of the backup.
          required: false
          type: string
      responses:
        '201':
          description: The Siddhi App is successfully restored.
          schema:
            $ref: '#/definitions/ApiResponseMessage'
        '409':
          description: The Siddhi App specified is not found.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
        '500':
          description: An unexpected error occured.
          schema:
            $ref: '#/definitions/ApiResponseMessageWithCode'
definitions:
  SiddhiAppContent:
    type: object
    properties:
      content:
        type: string
  SiddhiAppRevision:
    type: object
    properties:
      revision:
        type: string
  SiddhiAppStatus:
    type: object
    properties:
      status:
        type: string
  ApiResponseMessage:
    type: object
    properties:
      type:
        type: string
      message:
        type: string
  ApiResponseMessageWithCode:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string